+++ src/Translations.php
@@ -70,7 +70,7 @@ class Translations extends Plugin
     /**
      * @var string
      */
-    public $schemaVersion = '1.2.1';
+    public $schemaVersion = '1.2.2';
 
     // Public Methods
     // =========================================================================
@@ -335,6 +335,7 @@ class Translations extends Plugin
                     'translations/settings' => 'translations/settings/index',
                     'translations/settings/settings-check' => 'translations/settings/settings-check',
                     'translations/settings/send-logs' => 'translations/settings/send-logs',
+                    'translations/orders/get-file-diff/<fileId:\d+>' => 'translations/base/get-file-diff',
                 ]);
             }
         );