+++ src/migrations/Install.php
@@ -86,6 +86,7 @@ class Install extends Migration
                     'serviceFileId'     => $this->string()->defaultValue(''),
                     'dateCreated'       => $this->dateTime()->notNull(),
                     'dateUpdated'       => $this->dateTime()->notNull(),
+                    'dateDelivered'     => $this->dateTime(),
                     'uid'               => $this->uid()
                 ]
             );