+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/) and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 1.4.0 - 2019-10-18
+
+### Added
+- New source entries widget
+- Improved errors for failed API requests
+- Additional XML import error reporting
+- Bug fixes
+
+### Updated
+- Updated order detail page
+    - Compare file changes
+- API request methods
+
 ## 1.3.0 - 2019-10-18
 
 ### Added